tests: nightly: Fix and add more tests
Since cellular location method priority is no longer right after Wi-Fi,
the location library will not combine Wi-Fi and cellular methods.

Due to this we need to update the test to not expect a combined location
request.

Add additional checks to the same test to also verify that the device
samples other sensor sources.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>