power: Add timestamp to sampling response
Add timestamp to POWER_BATTERY_PERCENTAGE_SAMPLE_RESPONSE message
and send it to cloud.
This is done in preparation to storage module where samples
can be captured a long time in advance of sending the data.
Currently, the timestamp is added at the time of sending.

Rename test file from main.c to power_module_test.c to align with
other modules.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>