modules: Add storage module
The storage module adds the ability to buffer data in a chosen
backend. Currently, RAM backend is implemented.

The module subscribes to a list of zbus channels and stores all
the sampled data it receives on those channels, as long as there
is available space. The RAM buffer acts as a ring buffer, where
the oldest entry is discarded if all capacity is used.

There are two operational modes:
- Pass-through: Data is passed on immediately with no storage
- Buffer: Data is stored and passed on when the storage module
  receives a message over its zbus channel to do so.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>