modules: location: Cancel location request when
Cancel the location request if Wi-Fi and/or
neighbor cell information was sent to the cloud.

This is a workaround for a limitation in the location library
that otherwise requires the location to be set or it will move on
to the next fallback location method. As we do not
want the location to be sent back to the device, we cancel the
request to avoid that.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>