power: fix sign of current for SOC calculation
The Zephyr sensor API returns battery current as negative when
discharging and positive when charging[^1]. However, the nRF fuel gauge
library expects the opposite convention[^2]. When flipped, the SOC
calculation shows inverse charging behavior. This commit updates the
power module to flip the sign of the current value obtained from the
Zephyr sensor API before passing it to the fuel gauge lib.

[^1]: https://github.com/zephyrproject-rtos/zephyr/blob/7a65336edc345a81508303ae97caf8fa176a5a9f/include/zephyr/drivers/sensor.h#L172-L173
[^2]: https://github.com/nrfconnect/sdk-nrf/blob/270cebc0718dd035605dfb9ae2ec809c46c83fec/samples/pmic/native/npm13xx_fuel_gauge/src/fuel_gauge.c#L103-L106

Signed-off-by: Gillian Minnehan <gillian.minnehan@nordicsemi.no>